Update 7/12/2012
Re-enabled the Blizz Bar check to enable/disable it.
Add a new slash command /ion blizzbar to toggle the Blizzard Bars
Due to changes on Blizzard's side, it may no longer be possible to toggle the main bar and vehicle bar independently of each other, so the /ion blizzbar command gives you all or none.
Fixed an error for item scanning, related to the flyouts.
Item-based flyouts should also scan faster when updating.